% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/demography_functions.R
\name{modelmig}
\alias{modelmig}
\title{Estimate Model Migration Schedules using Non-linear Least Squares}
\usage{
modelmig(data, model, ages, profiles, initial = NULL)
}
\arguments{
\item{data}{Data frame with places in columns with an ID for age.}

\item{model}{Picks 7, 9, or 11 parameter models}

\item{ages}{Number of age categories, use the maximum age.}

\item{profiles}{Number of profiles/areas to fit.}
}
\value{
A factor, unless otherwise specified, with recoded \code{agevar}.
}
\description{
Estimate Model Migration Schedules using Non-linear Least Squares
}
\examples{
age.cat()
}
\keyword{Initital}
\keyword{Optional.}
\keyword{for}
\keyword{initital}
\keyword{model.}
\keyword{parameter}
\keyword{the}
\keyword{values}

