% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/graphic_functions.R
\name{relsize}
\alias{relsize}
\title{Calculates relative width and height for comparison from data.}
\usage{
relsize(comp, ref, h, w)
}
\arguments{
\item{comp}{Value of variable from comparison group.}

\item{ref}{Value of reference group, generally larger.}

\item{h}{Height of the reference square.}

\item{w}{Width of the reference square.}
}
\description{
Calculates relative width and height for comparison from data.
}
\keyword{annualized}
\keyword{growth}
\keyword{rate,}

