% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/export_functions.R
\name{save.xlsx}
\alias{save.xlsx}
\title{Exports R objects to a .xlsx file.}
\usage{
save.xlsx(file, ...)
}
\arguments{
\item{file}{Filename and necessary path to place new file.}

\item{...}{R objects to write to file (each to a new sheet).}
}
\description{
This function is wrapper function by Rob Kabacoff.
See: http://bit.ly/1wbMCe8.
It will save as many objects as you need to new sheets in Excel
}
\examples{
save.xlsx("example.xlsx", mtcars)
}
\keyword{Excel}
\keyword{export,}

