\name{zscore.rob}
\alias{zscore.rob}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Z-score Calculator
}
\description{Calculates z-score for numeric vector.
}
\usage{
zscore.rob(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Numeric vector.
}
}
\details{ }
\value{ }
\references{ }
\author{Robert S. Chavez}
\note{beta function}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{n/a}
\examples{
x <- rnorm(100,100,15)
zscore.rob(x)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{zscore.rob }
\keyword{z-score}